Conditional file download

+++ launch.sh
@@ -12,14 +12,21 @@ pip3 install -r flaskapp/requirements.txt
 echo -e "${GREEN}Installing spacy en model${NC}" 
 python -m spacy download en
 
-echo -e "${GREEN}Downloading required sentence embeddings${NC}" 
-gdown https://drive.google.com/uc?id=1l2liCZqWX3EfYpzv9OmVatJAEISPFihW
-unzip para-nmt-50m-demo.zip
-mv para-nmt-50m-demo data
-rm para-nmt-50m-demo.zip
+echo -e "${GREEN}Creating data directory${NC}" 
+mkdir -p data
+
+echo -e "${GREEN}Downloading required sentence embeddings${NC}"
+if [ ! -d "data/para-nmt-50m-demo" ] then 
+    gdown https://drive.google.com/uc?id=1l2liCZqWX3EfYpzv9OmVatJAEISPFihW
+    unzip para-nmt-50m-demo.zip
+    mv para-nmt-50m-demo data
+    rm para-nmt-50m-demo.zip
+fi
 
 echo -e "${GREEN}Downloading model checkpoint${NC}" 
-gdown https://drive.google.com/uc?id=18uOQsosF4uVGvUgp6pB4BKrQZ1FktlmM
-tar -xzf train-20180325-001253.tar.gz
-mv train-20180325-001253 data
-rm train-20180325-001253.tar.gz
+if [ ! -d "data/train-20180325-001253" ] then 
+    gdown https://drive.google.com/uc?id=18uOQsosF4uVGvUgp6pB4BKrQZ1FktlmM
+    tar -xzf train-20180325-001253.tar.gz
+    mv train-20180325-001253 data
+    rm train-20180325-001253.tar.gz
+fi
